Refactor: Change to service-serviceImpl structure
service-serviceImpl 구조로 변경
